requirements.txt
Flask
flask-cors
python-dotenv
PyMuPDF
pdf2image
pytesseract
openai